code cleanup: remove unused imports, add missing annotations
git-svn-id: file:///srv/local/data/cleanRepo/trunk@14421 a0517364-e873-4167-b1c8-416daa5a6b91

+++ matsim/src/main/java/org/matsim/locationchoice/facilityload/FacilityLoad.java
@@ -143,7 +143,7 @@ public class FacilityLoad {
 	public int timeBinIndex(double time) {
 		int lastBinIndex = this.numberOfTimeBins-1;
 		int numberOfBinsPerHour = this.numberOfTimeBins/24;
-		int secondsPerBin = (int)3600/numberOfBinsPerHour;
+		int secondsPerBin = 3600/numberOfBinsPerHour;
 		return Math.min(lastBinIndex, (int)(time/secondsPerBin));
 	}
 }
